meta-quanta: olympus-nuvoton: bmcweb Log Service
Fix the issue that the completion handler for one http request is executed many times.
This will cause a problem that subsequent http request might not be handled properly.

reproduce by:
1. POST /redfish/v1/Managers/bmc/LogServices/Dump/Actions/LogService.ClearLog
2. POST /redfish/v1/SessionService/Sessions
The second post request will not be handled correctly.

root cause:
The Response object is shared by many AsyncResp and its end() will be invoked
whenever the AsyncResp object is destroyed.

Signed-off-by: Stanley Chu <yschu@nuvoton.com>